[TokenCredentialStub] revert the introduction of credential stub as per pr feedback
